use int32_t

+++ protocol_buffers.cc
@@ -581,7 +581,7 @@ static const char* pb_decode_message(INTERNAL_FUNCTION_PARAMETERS, const char *d
                         break;
                         case TYPE_FIXED32:
                         {
-                            unsigned long l = 0;
+                            uint32_t l = 0;
                             memcpy(&l, data, 4);
 
                             MAKE_STD_ZVAL(dz);
@@ -627,7 +627,7 @@ static const char* pb_decode_message(INTERNAL_FUNCTION_PARAMETERS, const char *d
                         break;
                         case TYPE_SFIXED32:
                         {
-                            long l = 0;
+                            int32_t l = 0;
 
                             memcpy(&l, data, 4);
                             MAKE_STD_ZVAL(dz);
@@ -691,14 +691,14 @@ static const char* pb_decode_message(INTERNAL_FUNCTION_PARAMETERS, const char *d
                 MAKE_STD_ZVAL(dz);
                 ZVAL_DOUBLE(dz, a);
             } else if (s->type == TYPE_SFIXED32) {
-                long l = 0;
+                int32_t l = 0;
 
                 memcpy(&l, data, 4);
                 MAKE_STD_ZVAL(dz);
 
                 ZVAL_LONG(dz, (int32_t)l);
             } else if (s->type == TYPE_FIXED32) {
-                unsigned long l = 0;
+                uint32_t l = 0;
                 memcpy(&l, data, 4);
 
                 MAKE_STD_ZVAL(dz);
@@ -713,7 +713,7 @@ static const char* pb_decode_message(INTERNAL_FUNCTION_PARAMETERS, const char *d
                 ZVAL_LONG(dz, (unsigned long)l);
 #endif
             } else {
-                long l = 0;
+                int32_t l = 0;
 
                 memcpy(&l, data, 4);
